Renamed authoThunks into userThunks and added gymThunks
