+++ Frontend/src/screens/profileScreen.js
@@ -95,4 +95,5 @@ const styles = StyleSheet.create({
     fontWeight: 'bold',
   },
 });
+
 export default ProfileScreen;
